+++ numpy_turtle/__init__.py
@@ -1,1 +1,4 @@
-from numpy_turtle.numpy_turtle import Turtle
+from . import l_system
+from .numpy_turtle import Turtle
+
+__all__ = ["Turtle", "l_system"]